moving filter code to top level directory
