+++ roles/openshift_logging_elasticsearch/defaults/main.yml
@@ -6,8 +6,8 @@ openshift_logging_elasticsearch_image_pull_secret: "{{ openshift_hosted_logging_
 openshift_logging_elasticsearch_namespace: logging
 
 openshift_logging_elasticsearch_nodeselector: "{{ openshift_logging_es_nodeselector | default('') }}"
-openshift_logging_elasticsearch_cpu_limit: "{{ openshift_logging_es_cpu_limit | lower_bound_cpu('1000m') }}"
-openshift_logging_elasticsearch_memory_limit: "{{ openshift_logging_es_memory_limit | lower_bound_mem('1Gi') }}"
+openshift_logging_elasticsearch_cpu_limit: "{{ openshift_logging_es_cpu_limit | oo_lower_bound_cpu('1000m') }}"
+openshift_logging_elasticsearch_memory_limit: "{{ openshift_logging_es_memory_limit | oo_lower_bound_mem('1Gi') }}"
 openshift_logging_elasticsearch_recover_after_time: "{{ openshift_logging_es_recover_after_time | default('5m') }}"
 
 openshift_logging_elasticsearch_replica_count: 1
